Moved everything to one script
